refactor(patina): split script.rs inline tests into separate file
