+++ crates/vize_patina/src/rules/script/no_potential_component_option_typo/tests.rs
@@ -1,4 +1,3 @@
-
 use super::NoPotentialComponentOptionTypo;
 use crate::rules::script::ScriptLinter;
